get async working + tests working
