+++ packages/flutter/test/widgets/sensitive_content_error_handling_test.dart
@@ -20,87 +20,78 @@ void main() {
     );
   });
 
-  // testWidgets(
-  //   'when SensitiveContentService.setContentSensitivity fails, SensitiveContentHost.register does not update calculatedContentSensitivity',
-  //   (WidgetTester tester) async {
-  //     int setContentSensitivityCall = 0;
-  //     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
-  //       SystemChannels.sensitiveContent,
-  //       (MethodCall methodCall) async {
-  //         if (methodCall.method == 'SensitiveContent.setContentSensitivity') {
-  //           setContentSensitivityCall += 1;
-  //           if (setContentSensitivityCall == 1) {
-  //             // In the first call to set content sensitivity, throw exception to test
-  //             // SensitiveContentHost.register behavior.
-  //             throw Exception('test exception');
-  //           }
-  //         } else if (methodCall.method == 'SensitiveContent.getContentSensitivity') {
-  //           // The enum name for ContentSensitivity.autoSensitive. This will be the
-  //           // initial content sensitivity setting that we expect to persist when
-  //           // the call to set sensitive content sensitivity fails.
-  //           return 'autoSensitive';
-  //         } else if (methodCall.method == 'SensitiveContent.isSupported') {
-  //           return true;
-  //         }
-  //         return null;
-  //       },
-  //     );
+  testWidgets(
+    'when SensitiveContentService.setContentSensitivity fails, SensitiveContentHost.register throws FlutterError but still updates calculatedContentSensitivity',
+    (WidgetTester tester) async {
+      int setContentSensitivityCall = 0;
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
+        SystemChannels.sensitiveContent,
+        (MethodCall methodCall) async {
+          if (methodCall.method == 'SensitiveContent.setContentSensitivity') {
+            setContentSensitivityCall += 1;
+            if (setContentSensitivityCall == 1) {
+              // In the first call to set content sensitivity, throw exception to test
+              // SensitiveContentHost.register behavior.
+              throw Exception('test exception');
+            }
+          } else if (methodCall.method == 'SensitiveContent.getContentSensitivity') {
+            return 'autoSensitive';
+          } else if (methodCall.method == 'SensitiveContent.isSupported') {
+            return true;
+          }
+          return null;
+        },
+      );
 
-  //     await tester.pumpWidget(
-  //       SensitiveContent(sensitivity: ContentSensitivity.sensitive, child: Container()),
-  //     );
+      await tester.pumpWidget(
+        SensitiveContent(sensitivity: ContentSensitivity.sensitive, child: Container()),
+      );
 
-  //     expect(tester.takeException(), isA<FlutterError>());
-  //     expect(
-  //       sensitiveContentHost.calculatedContentSensitivity,
-  //       equals(ContentSensitivity.autoSensitive),
-  //     );
-  //   },
-  // );
-
-  // testWidgets(
-  //   'when SensitiveContentService.setContentSensitivity fails, SensitiveContentHost.unregister does not update calculatedContentSensitivity when no SensitiveContent widgets are left in the tree',
-  //   (WidgetTester tester) async {
-  //     int setContentSensitivityCall = 0;
-  //     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
-  //       SystemChannels.sensitiveContent,
-  //       (MethodCall methodCall) async {
-  //         if (methodCall.method == 'SensitiveContent.setContentSensitivity') {
-  //           setContentSensitivityCall += 1;
-  //           if (setContentSensitivityCall == 2) {
-  //             // In the second call to set content sensitivity, throw exception to test
-  //             // SensitiveContentHost.unregister behavior.
-  //             throw Exception('test exception');
-  //           }
-  //         } else if (methodCall.method == 'SensitiveContent.getContentSensitivity') {
-  //           // The enum name for ContentSensitivity.autoSensitive. This will be the
-  //           // content sensitivity setting before the sensitive SensitiveContent
-  //           // widget is built.
-  //           return 'autoSensitive';
-  //         } else if (methodCall.method == 'SensitiveContent.isSupported') {
-  //           return true;
-  //         }
-  //         return null;
-  //       },
-  //     );
-
-  //     await tester.pumpWidget(
-  //       SensitiveContent(sensitivity: ContentSensitivity.sensitive, child: Container()),
-  //     );
-  //     await tester.pumpWidget(Container());
-
-  //     expect(tester.takeException(), isA<FlutterError>());
-  //     expect(
-  //       sensitiveContentHost.calculatedContentSensitivity,
-  //       equals(ContentSensitivity.sensitive),
-  //     );
-  //   },
-  // );
+      expect(tester.takeException(), isA<FlutterError>());
+      expect(
+        sensitiveContentHost.calculatedContentSensitivity,
+        equals(ContentSensitivity.sensitive),
+      );
+    },
+  );
 
   testWidgets(
-    'when SensitiveContentService.setContentSensitivity fails, SensitiveContentHost.unregister does not update calculatedContentSensitivity when there are SensitiveContent widgets left in the tree',
+    'when SensitiveContentService.setContentSensitivity fails, SensitiveContentHost.unregister throws FlutterError when no SensitiveContent widgets are left in the tree',
     (WidgetTester tester) async {
-      // int setContentSensitivityCall = 0;
+      int setContentSensitivityCall = 0;
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
+        SystemChannels.sensitiveContent,
+        (MethodCall methodCall) async {
+          if (methodCall.method == 'SensitiveContent.setContentSensitivity') {
+            setContentSensitivityCall += 1;
+            if (setContentSensitivityCall == 2) {
+              // In the second call to set content sensitivity, throw exception to test
+              // SensitiveContentHost.unregister behavior.
+              throw Exception('test exception');
+            }
+          } else if (methodCall.method == 'SensitiveContent.getContentSensitivity') {
+            return 'autoSensitive';
+          } else if (methodCall.method == 'SensitiveContent.isSupported') {
+            return true;
+          }
+          return null;
+        },
+      );
+
+      await tester.pumpWidget(
+        SensitiveContent(sensitivity: ContentSensitivity.sensitive, child: Container()),
+      );
+      await tester.pumpWidget(Container());
+
+      expect(tester.takeException(), isA<FlutterError>());
+      expect(sensitiveContentHost.calculatedContentSensitivity, null);
+    },
+  );
+
+  testWidgets(
+    'when SensitiveContentService.setContentSensitivity fails, SensitiveContentHost.unregister throws FlutterError and still updates calculatedContentSensitivity when there are SensitiveContent widgets left in the tree',
+    (WidgetTester tester) async {
+      int setContentSensitivityCall = 0;
       const Key scKey = Key('sc');
       final DisposeTester sc = DisposeTester(
         child: SensitiveContent(
@@ -118,19 +109,13 @@ void main() {
         SystemChannels.sensitiveContent,
         (MethodCall methodCall) async {
           if (methodCall.method == 'SensitiveContent.setContentSensitivity') {
-            // print('hello?@?!?!?!?!');
-            // setContentSensitivityCall += 1;
-            // // print('setContent called ${methodCall.arguments}, $setContentSensitivityCall');
-            // if (setContentSensitivityCall == 3) {
-            //   print('throwing exception');
-            //   // In the second call to set content sensitivity, throw exception to test
-            //   // SensitiveContentHost.unregister behavior.
-            //   throw Exception('test exception');
-            // }
+            setContentSensitivityCall += 1;
+            if (setContentSensitivityCall == 2) {
+              // In the second call to set content sensitivity, throw exception to test
+              // SensitiveContentHost.unregister behavior.
+              throw Exception('test exception');
+            }
           } else if (methodCall.method == 'SensitiveContent.getContentSensitivity') {
-            // The enum name for ContentSensitivity.autoSensitive. This will be the
-            // content sensitivity setting before the sensitive SensitiveContent
-            // widget is built.
             return 'autoSensitive';
           } else if (methodCall.method == 'SensitiveContent.isSupported') {
             return true;
@@ -139,19 +124,25 @@ void main() {
         },
       );
 
-      await tester.pumpWidget(Column(children: <Widget>[sc, asc]));
+      await tester.runAsync(() async {
+        await tester.pumpWidget(Column(children: <Widget>[sc, asc]));
 
-      final DisposeTesterState scDiposeTesterState = tester.firstState<DisposeTesterState>(
-        find.byKey(scKey),
-      );
-      scDiposeTesterState.disposeWidget();
-      await tester.pump();
+        final DisposeTesterState scDiposeTesterState = tester.firstState<DisposeTesterState>(
+          find.byKey(scKey),
+        );
+        scDiposeTesterState.disposeWidget();
 
-      expect(tester.takeException(), isA<FlutterError>());
-      expect(
-        sensitiveContentHost.calculatedContentSensitivity,
-        equals(ContentSensitivity.sensitive),
-      );
+        // Delay added to ensure that the SensitiveContent widget unregister completes.
+        await Future<void>.delayed(const Duration(milliseconds: 100), () async {
+          await tester.pumpAndSettle();
+        });
+
+        expect(tester.takeException(), isA<FlutterError>());
+        expect(
+          sensitiveContentHost.calculatedContentSensitivity,
+          equals(ContentSensitivity.autoSensitive),
+        );
+      });
     },
   );
 }
